+++ SalaoVirtual/FornecedorProdutos.csv
@@ -1,0 +1,2 @@
+1;1;10.3
+1;2;15.0
