+++ SalaoVirtual/nbproject/private/private.xml
@@ -2,6 +2,12 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/C:/Users/rafae/Desktop/Unicamp/SI%20-%204º%20Semestre/POO%20II/SalaoVirtual/SalaoVirtual/src/salaovirtual/SalaoVirtual.java</file>
+            <file>file:/C:/Users/rafae/Desktop/Unicamp/SI%20-%204º%20Semestre/POO%20II/SalaoVirtual/SalaoVirtual/src/salaovirtual/Cadastro.java</file>
+            <file>file:/C:/Users/rafae/Desktop/Unicamp/SI%20-%204º%20Semestre/POO%20II/SalaoVirtual/SalaoVirtual/src/salaovirtual/Servico.java</file>
+            <file>file:/C:/Users/rafae/Desktop/Unicamp/SI%20-%204º%20Semestre/POO%20II/SalaoVirtual/SalaoVirtual/src/salaovirtual/Consulta.java</file>
+            <file>file:/C:/Users/rafae/Desktop/Unicamp/SI%20-%204º%20Semestre/POO%20II/SalaoVirtual/SalaoVirtual/src/salaovirtual/Fornecedor.java</file>
+        </group>
     </open-files>
 </project-private>
