+++ SalaoVirtual/src/salaovirtual/Servico.java
@@ -1,30 +1,19 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Classe referente ao serviço
  */
 package salaovirtual;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import static java.lang.Float.parseFloat;
-import static java.lang.Integer.parseInt;
-import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 
 /**
- *
- * @author r176257
+ * Classe referente ao fornecedor
+ * 
+ * Atributos que merecem destaque para explicação:
+ * estado: É o estado do serviço, podendo ser apenas "Agendado", "Cancelado" ou "Realizado"
+ * 
+ * @author Rafael Tavares
  */
 public class Servico {
     private int codigo;
@@ -35,13 +24,17 @@ public class Servico {
     private Funcionario funcionario;
     private Cliente cliente;
     
-    // Criar método para listar todos os serviços de um mês
-    // Criar método para listar todos os serviços de um ano
-    // Criar método para listar todos os serviços que um cliente fez
-    // Criar método para listar todos os serviços que um funcionário fez
     // A data está armazenando apena dd/MM/yyyy , precisa fazer armazenar horário também!
     
-
+    /**
+     * Agenda o serviço que chamou este método para (respectivamente aos parâmetros) um cliente e 
+     * um funcionário numa data especificada. 
+     * A data deve ser posteiror ao momento atual, caso seja anterior ocorrerá a DataInvalidaException.
+     * @param cli
+     * @param fun
+     * @param data
+     * @throws DataInvalidaException 
+     */
     public void agendarServico(Cliente cli, Funcionario fun, Date data) throws DataInvalidaException {
         Date agora = new Date();
         if (data.after(agora)) {   // Se for depois de agora, então é um agendamento válido
@@ -51,7 +44,7 @@ public class Servico {
             try {
                 this.setEstado("Agendado");
             } catch (EstadoServicoInvalidoException ex) {
-                Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
+                //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
             }
         }
         else {
@@ -59,12 +52,15 @@ public class Servico {
         }
     }
     
+    /**
+     * Efetua o serviço que chamou este método apenas se seu estado for Agendado ou nulo (o que significa que
+     * o serviço foi efetuado sem um agendamento prévio)
+     */
     public void efetuarServico() {
         try {
             if (this.getEstado().equals("Agendado")){
                 try {
                     this.setEstado("Realizado");
-                    gravarServico();
                 }
                 catch (EstadoServicoInvalidoException e) {
                     // Mandar mensagem de "O serviço não pôde ser realizado, favor contatar o desenvolvedor do sistema"
@@ -78,7 +74,6 @@ public class Servico {
                 Date data = new Date();
                 this.setEstado("Realizado");
                 this.setData(data);
-                gravarServico();
             }
             catch (EstadoServicoInvalidoException ex) {
                 // Mandar mensagem de "O serviço não pôde ser realizado, favor contatar o desenvolvedor do sistema"
@@ -86,6 +81,9 @@ public class Servico {
         }
     }
     
+    /**
+     * Modifica o estado do serviço que chamou este método para Cancelado.
+     */
     public void cancelarServico() {
         try {
             this.setEstado("Cancelado");
@@ -96,6 +94,11 @@ public class Servico {
         }
     }
     
+    /**
+     * Método para facilitar a escrita do objeto em um arquivo CSV
+     * Polimorfismo: Sobrescrita
+     * @return Atributos do objeto separados por ;
+     */
     @Override
     public String toString() {
         SimpleDateFormat formato = new SimpleDateFormat("dd/MM/yyyy");
@@ -103,253 +106,135 @@ public class Servico {
                     this.estado + ";" + this.funcionario.getLogin() + ";" + this.cliente.getCodigo();
     }
     
-    private void gravarServico() {
-        try {
-            FileWriter arq = new FileWriter("Servico.csv", true);
-            BufferedWriter saida = new BufferedWriter(arq);
-            this.setCodigo(this.getProxCodigo());
-            saida.write(this.toString());
-            saida.newLine();
-            saida.close();
-            arq.close();
-        } catch (IOException e) {
-            //
-        }
-    }
-    
-    public Servico encontrarServico(int codigo) {
-        try {
-            FileReader arq = new FileReader("Servico.csv");
-            BufferedReader entrada = new BufferedReader(arq);
-            String linha;
-            Servico s = new Servico();
-            SimpleDateFormat formato = new SimpleDateFormat("dd/MM/yyyy");
-            do {
-                linha = entrada.readLine();
-                String[] valor = linha.split(";");
-                if (parseInt(valor[0]) == codigo) {
-                    s.setCodigo(parseInt(valor[0]));
-                    s.setNome(valor[1]);
-                    s.setValor(parseFloat(valor[2]));
-                    s.setData(formato.parse(valor[3]));
-                    try {
-                        s.setEstado(valor[4]);
-                    } catch (EstadoServicoInvalidoException ex) {
-                        //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-                    }
-                    
-                    Funcionario f = new Funcionario();
-                    s.setFuncionario(f.encontrarFuncionarioLogin(valor[5]));
-                    
-                    Cliente c = new Cliente();
-                    s.setCliente(c.encontrarCliente(parseInt(valor[6])));
-                    
-                    entrada.close();
-                    arq.close();
-                    return s;
-                }
-            } while (linha != null);
-            entrada.close();
-            arq.close();
-        } catch (FileNotFoundException ex) {
-            return null;
-        }catch (IOException ex2) {
-            return null;
-        } catch (ParseException ex) {
-            //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return null;
-    }
-    
-    public List<Servico> encontrarServicoNome(String nome) {
-        try {
-            List<Servico> servicos = new ArrayList();
-            FileReader arq = new FileReader("Servico.csv");
-            BufferedReader entrada = new BufferedReader(arq);
-            String linha;
-            
-            linha = entrada.readLine();
-            SimpleDateFormat formato = new SimpleDateFormat("dd/MM/yyyy");
-            do {
-                Servico s = new Servico();
-                String[] valor = linha.split(";");
-                if (valor[1].equals(nome)) {
-                    s.setCodigo(parseInt(valor[0]));
-                    s.setNome(valor[1]);
-                    s.setValor(parseFloat(valor[2]));
-                    s.setData(formato.parse(valor[3]));
-                    try {
-                        s.setEstado(valor[4]);
-                    } catch (EstadoServicoInvalidoException ex) {
-                        //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-                    }
-                    
-                    Funcionario f = new Funcionario();
-                    s.setFuncionario(f.encontrarFuncionarioLogin(valor[5]));
-                    
-                    Cliente c = new Cliente();
-                    s.setCliente(c.encontrarCliente(parseInt(valor[6])));
-                    servicos.add(s);
-                }
-                linha = entrada.readLine();
-            } while (linha != null);
-            entrada.close();
-            arq.close();
-            return servicos;
-        } catch (FileNotFoundException e) {
-            return null;
-        } catch (IOException ex2) {
-            return null;
-        } catch (ParseException ex) {
-            //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return null;
-    }
-    
-    public List<Servico> encontrarServicoEstado(String estado) {
-        try {
-            Servico stemp = new Servico();
-            stemp.setEstado(estado);
-        } catch (EstadoServicoInvalidoException ex) {
-            return null;
-        }
-        
-        try {
-            List<Servico> servicos = new ArrayList();
-            FileReader arq = new FileReader("Servico.csv");
-            BufferedReader entrada = new BufferedReader(arq);
-            String linha;
-            
-            linha = entrada.readLine();
-            SimpleDateFormat formato = new SimpleDateFormat("dd/MM/yyyy");
-            do {
-                Servico s = new Servico();
-                String[] valor = linha.split(";");
-                if (valor[4].equals(estado)) {
-                    s.setCodigo(parseInt(valor[0]));
-                    s.setNome(valor[1]);
-                    s.setValor(parseFloat(valor[2]));
-                    s.setData(formato.parse(valor[3]));
-                    try {
-                        s.setEstado(valor[4]);
-                    } catch (EstadoServicoInvalidoException ex) {
-                        //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-                    }
-                    
-                    Funcionario f = new Funcionario();
-                    s.setFuncionario(f.encontrarFuncionarioLogin(valor[5]));
-                    
-                    Cliente c = new Cliente();
-                    s.setCliente(c.encontrarCliente(parseInt(valor[6])));
-                    servicos.add(s);
-                }
-                linha = entrada.readLine();
-            } while (linha != null);
-            entrada.close();
-            arq.close();
-            return servicos;
-        } catch (FileNotFoundException e) {
-            return null;
-        } catch (IOException ex2) {
-            return null;
-        } catch (ParseException ex) {
-            //Logger.getLogger(Servico.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        return null;
-    }
-    
-    public int getProxCodigo() {
-        int cod = 1;
-        try {
-            FileReader arq = new FileReader("Servico.csv");
-            BufferedReader entrada = new BufferedReader(arq);
-            String linha;
-            linha = entrada.readLine();
-            while (linha != null) {
-                cod++;
-                linha = entrada.readLine();
-            }
-            entrada.close();
-            arq.close();
-        } catch (FileNotFoundException ex) {
-            try {
-                FileWriter arq = new FileWriter("Servico.csv");
-                BufferedWriter saida = new BufferedWriter(arq);
-                saida.close();
-                arq.close();
-            } catch (IOException ex1) {
-                //Logger.getLogger(Cliente.class.getName()).log(Level.SEVERE, null, ex1);
-            }
-        } catch (IOException ex2) {
-            // log
-        }
-        return cod;
-    }
-    
     /* Métodos Construtores, Getters & Setters */
+    /**
+     * Método construtor para facilitar a criação de um objeto que será cadastrado no sistema
+     * @param nome 
+     */
     public Servico(String nome) {
         this.nome = nome;
     }
 
+    /**
+     * Método construtor para facilitar a criação de um objeto que será utilizado para consulta ao invés
+     * de cadastro
+     */
     public Servico() {
         
     }
 
-    private void setCodigo(int codigo) {
+    /**
+     * Define o código do serviço
+     * @param codigo 
+     */
+    public void setCodigo(int codigo) {
         this.codigo = codigo;
     }
 
+    /**
+     * Define o nome do serviço
+     * @param nome 
+     */
     public void setNome(String nome) {
         this.nome = nome;
     }
 
+    /**
+     * Retorna o valor do serviço
+     * @return Valor
+     */
     public float getValor() {
         return valor;
     }
 
+    /**
+     * Define o valor do serviço
+     * @param valor 
+     */
     public void setValor(float valor) {
         this.valor = valor;
     }
     
+    /**
+     * Retorna o nome do serviço
+     * @return Nome
+     */
     public String getNome() {
         return nome;
     }
 
+    /**
+     * Retorna o código do serviço
+     * @return Código
+     */
     public int getCodigo() {
         return codigo;
     }
 
+    /**
+     * Retorna a data do serviço
+     * @return Data
+     */
     public Date getData() {
         return data;
     }
 
-    private void setData(Date data) {
+    /**
+     * Define a data do serviço
+     * @param data 
+     */
+    public void setData(Date data) {
         this.data = data;
     }
 
+    /**
+     * Retorna o estado atual do serviço
+     * @return Estado
+     */
     public String getEstado() {
         return estado;
     }
 
-    private void setEstado(String estado) throws EstadoServicoInvalidoException {
+    /**
+     * Define o estado atual do serviço. Gera EstadoServicoInvalidoException caso não seja um estado válido
+     * @param estado
+     * @throws EstadoServicoInvalidoException 
+     */
+    public void setEstado(String estado) throws EstadoServicoInvalidoException {
         if ((estado.equals("Agendado")) || (estado.equals("Realizado")) || (estado.equals("Cancelado")))
             this.estado = estado;
         else
             throw new EstadoServicoInvalidoException();
     }
 
+    /**
+     * Retorna o funcionário responsável por este serviço
+     * @return Funcionário
+     */
     public Funcionario getFuncionario() {
         return funcionario;
     }
 
+    /**
+     * Define o funcionário responsável por este serviço
+     * @param funcionario 
+     */
     public void setFuncionario(Funcionario funcionario) {
         this.funcionario = funcionario;
     }
 
+    /**
+     * Retorna o cliente que recebeu este serviço
+     * @return 
+     */
     public Cliente getCliente() {
         return cliente;
     }
 
+    /**
+     * Define o cliente que recebeu este serviço
+     * @param cliente 
+     */
     public void setCliente(Cliente cliente) {
         this.cliente = cliente;
     }
